Update weight_matching.py
dot not implemented in half error

+++ weight_matching.py
@@ -813,10 +813,8 @@ def weight_matching(ps: PermutationSpec, params_a, params_b, max_iter=1, init_pe
 
           assert (torch.tensor(ri) == torch.arange(len(ri))).all()
           
-          oldL = torch.vdot(torch.flatten(A).float(), torch.flatten(torch.eye(n)[perm[p].long()]))
-          newL = torch.vdot(torch.flatten(A).float(), torch.flatten(torch.eye(n)[ci, :]))
-          oldL = oldL.half()
-          newL = newL.half()
+          oldL = torch.vdot(torch.flatten(A).float(), torch.flatten(torch.eye(n)[perm[p].long()]).float()).half()
+          newL = torch.vdot(torch.flatten(A).float(), torch.flatten(torch.eye(n)[ci, :]).float()).half()
           
           if newL - oldL != 0:
             sum += abs((newL-oldL).item())
